refactor: move isLookaheadRelational to flow plugins
